Moved ReceivedMessagesBuffer to seperate module
